split test helpers into a new filename
